+++ controller/editorcontroller.php
@@ -1,7 +1,7 @@
 <?php
 /**
  *
- * (c) Copyright Ascensio System SIA 2021
+ * (c) Copyright Ascensio System SIA 2022
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -200,6 +200,7 @@ class EditorController extends Controller {
      * @param string $name - file name
      * @param string $dir - folder path
      * @param string $templateId - file identifier
+     * @param int $targetId - identifier of the file for using as template for create
      * @param string $shareToken - access token
      *
      * @return array
@@ -207,7 +208,7 @@ class EditorController extends Controller {
      * @NoAdminRequired
      * @PublicPage
      */
-    public function create($name, $dir, $templateId = null, $shareToken = null) {
+    public function create($name, $dir, $templateId = null, $targetId = 0, $shareToken = null) {
         $this->logger->debug("Create: $name", ["app" => $this->appName]);
 
         if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
@@ -219,8 +220,10 @@ class EditorController extends Controller {
             return ["error" => $this->trans->t("Template not found")];
         }
 
+        $user = null;
         if (empty($shareToken)) {
-            $userId = $this->userSession->getUser()->getUID();
+            $user = $this->userSession->getUser();
+            $userId = $user->getUID();
             $userFolder = $this->root->getUserFolder($userId);
         } else {
             list ($userFolder, $error, $share) = $this->fileUtility->getNodeByToken($shareToken);
@@ -251,13 +254,30 @@ class EditorController extends Controller {
             return ["error" => $this->trans->t("You don't have enough permission to create")];
         }
 
-        if (empty($templateId)) {
-            $template = TemplateManager::GetEmptyTemplate($name);
-        } else {
+        if (!empty($templateId)) {
             $templateFile = TemplateManager::GetTemplate($templateId);
             if ($templateFile !== null) {
                 $template = $templateFile->getContent();
             }
+        } elseif (!empty($targetId)) {
+            $targetFile = $userFolder->getById($targetId)[0];
+            $targetName = $targetFile->getName();
+            $targetExt = strtolower(pathinfo($targetName, PATHINFO_EXTENSION));
+            $targetKey = $this->fileUtility->getKey($targetFile);
+            
+            $fileUrl = $this->getUrl($targetFile, $user, $shareToken);
+
+            $ext = strtolower(pathinfo($name, PATHINFO_EXTENSION));
+            $documentService = new DocumentService($this->trans, $this->config);
+            try {
+                $newFileUri = $documentService->GetConvertedUri($fileUrl, $targetExt, $ext, $targetKey);
+            } catch (\Exception $e) {
+                $this->logger->logException($e, ["message" => "GetConvertedUri: " . $targetFile->getId(), "app" => $this->appName]);
+                return ["error" => $e->getMessage()];
+            }
+            $template = $documentService->Request($newFileUri);
+        } else {
+            $template = TemplateManager::GetEmptyTemplate($name);
         }
 
         if (!$template) {
@@ -671,16 +691,15 @@ class EditorController extends Controller {
      * Get versions history for file
      *
      * @param integer $fileId - file identifier
-     * @param string $shareToken - access token
      *
      * @return array
      *
      * @NoAdminRequired
      */
-    public function history($fileId, $shareToken = null) {
+    public function history($fileId) {
         $this->logger->debug("Request history for: $fileId", ["app" => $this->appName]);
 
-        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
+        if (!$this->config->isUserAllowedToUse()) {
             return ["error" => $this->trans->t("Not permitted")];
         }
 
@@ -692,7 +711,7 @@ class EditorController extends Controller {
             $userId = $user->getUID();
         }
 
-        list ($file, $error, $share) = empty($shareToken) ? $this->getFile($userId, $fileId) : $this->fileUtility->getFileByToken($fileId, $shareToken);
+        list ($file, $error, $share) = $this->getFile($userId, $fileId);
 
         if (isset($error)) {
             $this->logger->error("History: $fileId $error", ["app" => $this->appName]);
@@ -791,16 +810,15 @@ class EditorController extends Controller {
      *
      * @param integer $fileId - file identifier
      * @param integer $version - file version
-     * @param string $shareToken - access token
      *
      * @return array
      *
      * @NoAdminRequired
      */
-    public function version($fileId, $version, $shareToken = null) {
+    public function version($fileId, $version) {
         $this->logger->debug("Request version for: $fileId ($version)", ["app" => $this->appName]);
 
-        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
+        if (!$this->config->isUserAllowedToUse()) {
             return ["error" => $this->trans->t("Not permitted")];
         }
 
@@ -812,7 +830,7 @@ class EditorController extends Controller {
             $userId = $user->getUID();
         }
 
-        list ($file, $error, $share) = empty($shareToken) ? $this->getFile($userId, $fileId) : $this->fileUtility->getFileByToken($fileId, $shareToken);
+        list ($file, $error, $share) = $this->getFile($userId, $fileId);
 
         if (isset($error)) {
             $this->logger->error("History: $fileId $error", ["app" => $this->appName]);
@@ -841,14 +859,14 @@ class EditorController extends Controller {
             $key = $this->fileUtility->getKey($file, true);
             $versionId = $file->getFileInfo()->getMtime();
 
-            $fileUrl = $this->getUrl($file, $user, $shareToken);
+            $fileUrl = $this->getUrl($file, $user);
         } else {
             $fileVersion = array_values($versions)[$version - 1];
 
             $key = $this->fileUtility->getVersionKey($fileVersion);
             $versionId = $fileVersion->getRevisionId();
 
-            $fileUrl = $this->getUrl($file, $user, $shareToken, $version);
+            $fileUrl = $this->getUrl($file, $user, null, $version);
         }
         $key = DocumentService::GenerateRevisionId($key);
 
@@ -862,14 +880,14 @@ class EditorController extends Controller {
             && count($versions) >= $version - 1
             && FileVersions::hasChanges($ownerId, $fileId, $versionId)) {
 
-            $changesUrl = $this->getUrl($file, $user, $shareToken, $version, true);
+            $changesUrl = $this->getUrl($file, $user, null, $version, true);
             $result["changesUrl"] = $changesUrl;
 
             $prevVersion = array_values($versions)[$version - 2];
             $prevVersionKey = $this->fileUtility->getVersionKey($prevVersion);
             $prevVersionKey = DocumentService::GenerateRevisionId($prevVersionKey);
 
-            $prevVersionUrl = $this->getUrl($file, $user, $shareToken, $version - 1);
+            $prevVersionUrl = $this->getUrl($file, $user, null, $version - 1);
 
             $result["previous"] = [
                 "key" => $prevVersionKey,
@@ -890,17 +908,16 @@ class EditorController extends Controller {
      *
      * @param integer $fileId - file identifier
      * @param integer $version - file version
-     * @param string $shareToken - access token
      *
      * @return array
      *
      * @NoAdminRequired
      * @PublicPage
      */
-    public function restore($fileId, $version, $shareToken = null) {
+    public function restore($fileId, $version) {
         $this->logger->debug("Request restore version for: $fileId ($version)", ["app" => $this->appName]);
 
-        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
+        if (!$this->config->isUserAllowedToUse()) {
             return ["error" => $this->trans->t("Not permitted")];
         }
 
@@ -912,7 +929,7 @@ class EditorController extends Controller {
             $userId = $user->getUID();
         }
 
-        list ($file, $error, $share) = empty($shareToken) ? $this->getFile($userId, $fileId) : $this->fileUtility->getFileByToken($fileId, $shareToken);
+        list ($file, $error, $share) = $this->getFile($userId, $fileId);
 
         if (isset($error)) {
             $this->logger->error("Restore: $fileId $error", ["app" => $this->appName]);
@@ -937,7 +954,7 @@ class EditorController extends Controller {
             }
         }
 
-        return $this->history($fileId, $shareToken);
+        return $this->history($fileId);
     }
 
     /**